Add ability to edit quests
Tapping on the quest card border on the Quest List page, when no quests are selected, brings the player to the New Quest page and lets them edit the selected quest's data.

Editing quests is only allowed when no quests are selected to prevent the player from accidentally tapping the card border when trying to select multiple quests.