Stop updating an edited quest's create date
The app re-introduces a bug after the quest save was moved to the New Quest view model.

The bug is triggered by editing a quest, which updates its creation date even though that value should not change after the quest is created.

This fixes the bug by referencing the database record and only updating the name, difficulty and icon.