Implement moving icons to a different icon group
Adds the ability to select multiple icons and move them to a different icon group. Makes the use of Edit and Move modes in Icon Select.

Edit mode allows player to select multiple icons and act on them. Edit mode also changes the Add New Icon button to the Move button.

Clicking on the Move button triggers Move mode, which changes the Move button to a Cancel button and which disables all buttons except the icon group buttons and the cancel button.

In Move mode, the icon group buttons enlarge and move towards the center of the screen to make it easier to select one. Selecting an icon group moves selected icons to that icon group.

Pressing the cancel button exits Move mode back to Edit mode and reverts buttons to their previous state before Move mode was triggered.