Fix icon group buttons getting misaligned sometimes
If users spam the move and cancel button in Icon Select, they can easily get the icon group buttons to get stuck in what looks like partway through the animation.

This is caused when running the zoom-in position calculations while the buttons are mid-animation, which throws the calculations off.

Fix this by making the zoom-in calculations rely on the button's original positions and not their current ones.