Animate icons fading out after moving
To show feedback that the icons were actually moved.

Originally, the user has to switch to a different icon group and back to show that the icons actually changed.

Modifies Icon Select to load icon lists from the database once, then maintain icon lists on frontend manually on top of updating the database.