Prevent input field jumps on confirm
When there are multiple EditTexts on the same page, confirming on one higher on the screen does not close the keyboard and instead jumps to the next input field awaiting input.

These input jumps do not work will for this app since only a few fields have to be changed at the time most of the time.

Preventing these jumps improves app flow.