Fix user Settings input not getting cleared
If the user presses the back button, then user input saved isn't cleared, unlike when the other navigation options are used.

This update customizes the back button in the Settings page to clear saved user input before navigating away from the page, fixing the bug.

Also refactors some code to align with the rest of the codebase or to make things more concise.