Fix Edit Quest data not clearing on back nav
Using built-in back navigation on Edit Quest resulted in input not getting cleared. As a result, trying to create a new quest afterwards caused the Edit Quest page to appear instead.

This bug is fixed by clearing Edit Quest saved data from the view model whenever the built-in back navigation is triggered on the page.

Also refactors similar code used on the Settings page so that it is more reusable.