Hide Shop sort by default and reveal on tap
Add an invisible trigger zone between the bottom two corner buttons in the Shop that, when tapped, reveal the sort button.

After a period of time, the sort button hides itself.

Interacting with the sort button will make it stay on screen for longer.

This way the user only needs to reveal the sort button when they need it, which probably isn't that often.