Fix sort button trigger blocking quests and items
The invisible trigger that causes the sort button to appear sometimes blocks users from selecting quests and items near the bottom of their screen, forcing the user to scroll a bit or tap around the trigger.

This moves the trigger behind the cards/items and to the sides of the Quest List and Shop so that the trigger never blocks quests/items.