Fix card colors and icons not changing on edit
Fixes an issue where editing a card's difficulty does not change the card's color in Quest List until you move to a different screen and back.

It turns out compose was not updating the background color since it didn't have a key to tell when to update the color or not.